autotest: moving test lib as a clean dual C11/Rust implementation
Rust implementation of test suites to be added in another PR
