+++ autotest/include/meson.build
@@ -1,11 +1,4 @@
-# SPDX-FileCopyrightText: 2023 - 2025 Ledger SAS
+# SPDX-FileCopyrightText: 2025 H2Lab OSS Team
 # SPDX-License-Identifier: Apache-2.0
 
 autotest_inc = include_directories('.')
-
-autotest_headerset.add(
-    files(
-        'testlib/assert.h',
-        'testlib/log.h',
-    )
-)
