+++ autotest/src/tests/include/devices-dt.h.in
