+++ autotest/src/tests/include/test_map.h
