+++ autotest/src/tests/src/test_cycles.c
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include <inttypes.h>
-#include <testlib/log.h>
-#include <testlib/assert.h>
+#include <test_log.h>
+#include <test_assert.h>
 #include <uapi/uapi.h>
 #include "test_cycles.h"
